+++ ui.c
@@ -38,9 +38,10 @@ double min_freq	= DEFAULT_MIN_FREQ;
 double max_freq	= DEFAULT_MAX_FREQ;	
 
 /* Dynamic range of color map (values below this are black) */
-double dyn_range	= DEFAULT_DYN_RANGE;
+float dyn_range = DEFAULT_DYN_RANGE;
+
 /* Highest value seen so far in spectrogram data. */
-double logmax = DEFAULT_LOGMAX;
+float logmax = DEFAULT_LOGMAX;
 
 /* How many video output frames to generate per second while playing
  * and how many pixel columns to generate per second of the audio file.
